--- SwiftUIs/SwiftUIs/Learning/Project 2/GridStack.swift
@@ -1,33 +1,0 @@
-//
-//  GridStack.swift
-//  SwiftUIExamples
-//
-//  Created by Cedric Bahirwe on 8/28/20.
-//  Copyright © 2020 Cedric Bahirwe. All rights reserved.
-//
-
-import SwiftUI
-
-struct GridStack<Content: View>: View {
-    let rows: Int
-    let columns: Int
-    let content: (Int, Int) -> Content
-    var body: some View {
-        VStack {
-            ForEach(0 ..< self.rows) { row in
-                HStack {
-                    ForEach(0 ..<  self.columns) { column in
-                        self.content(row, column)
-                            .padding(.horizontal, 10)
-                            .padding(.vertical, 5)
-                    }
-                }
-            }
-        }
-    }
-    init(rows: Int, columns: Int, @ViewBuilder content: @escaping (Int, Int) -> Content ) {
-        self.rows = rows
-        self.columns = columns
-        self.content = content
-    }
-}
